+++ Common/Melee/_Animations/QuickSlashMeleeAnimation.cs
@@ -28,27 +28,27 @@ public class QuickSlashMeleeAnimation : MeleeAnimation, ICanDoMeleeDamage, IModi
 			baseAngle = 0f;
 		}
 
+		if (item.TryGetGlobalItem(out ItemPowerAttacks powerAttacks) && powerAttacks.Enabled && powerAttacks.IsCharging) {
+			baseAngle -= MathHelper.PiOver2 * 0.25f * player.direction;
+		}
+
 		float step = 1f - MathHelper.Clamp(player.itemAnimation / (float)player.itemAnimationMax, 0f, 1f);
 		int dir = player.direction * (IsAttackFlipped ? -1 : 1);
 
-		float minValue = baseAngle - MathHelper.PiOver2 * 1.25f;
-		float maxValue = baseAngle + MathHelper.PiOver2 * 1.0f;
-
-		if (dir < 0) {
-			Utils.Swap(ref minValue, ref maxValue);
-		}
+		float minValue = baseAngle - (MathHelper.PiOver2 * 1.25f * dir);
+		float maxValue = baseAngle + (MathHelper.PiOver2 * 1.00f * dir);
 
 		var animation = new Gradient<float>(
-			(0.0f, minValue),
-			(0.1f, minValue),
-			(0.15f, MathHelper.Lerp(minValue, maxValue, 0.125f)),
-			(0.151f, MathHelper.Lerp(minValue, maxValue, 0.8f)),
-			(0.5f, maxValue),
-			(0.8f, MathHelper.Lerp(minValue, maxValue, 0.8f)),
-			(1.0f, MathHelper.Lerp(minValue, maxValue, 0.8f))
+			(0.000f, 0.000f),
+			(0.100f, 0.000f),
+			(0.150f, 0.125f),
+			(0.151f, 0.800f),
+			(0.500f, 1.000f),
+			(0.800f, 0.800f),
+			(1.000f, 0.800f)
 		);
 
-		return animation.GetValue(step);
+		return MathHelper.Lerp(minValue, maxValue, animation.GetValue(step));
 	}
 
 	// Direction switching